Refactor + fixes for windows : skip file when finding latest version + fixed error
